+++ packages/stream_video_flutter/CHANGELOG.md
@@ -20,6 +20,7 @@ By (only) using these callbacks the root widgets will use more efficient partial
 🐞 Fixed
 * Improved SFU error handling in Call flow and disconnect reason handling. The disconnected call state now accurately reflects the original cause of disconnection.
 * Fixed an issue where rejecting a ringing call by one participant would incorrectly end the call for all other ringing participants.
+* Fixed an edge case where a call with the same CID as an incoming call is also an outgoing call to ensure the same Call instance is used.
 
 ## 0.9.6